Install puppet from the Ubuntu apt repository for aarch64/arm64
The Puppet Labs apt repository does not provide packages for arm64 architecture, preventing
this script from installing puppet on arm64 systems running Ubuntu.  This prevents the use
of Nodepool and the nodepool scripts in openstack-infra/project-config from being used unmodified.

This change determines the host architecture, and skips the direct download of the puppet .deb
from apt.puppetlabs.com, instead just installing puppet directly from the Ubuntu apt
repository.

+++ install_puppet.sh
@@ -188,10 +188,15 @@ Pin: version $FACTER_VERSION
 Pin-Priority: 501
 EOF
 
-    puppet_deb=puppetlabs-release-${lsbdistcodename}.deb
-    wget http://apt.puppetlabs.com/$puppet_deb -O $puppet_deb
-    dpkg -i $puppet_deb
-    rm $puppet_deb
+    # if running on an architecture (currently aarch64/arm64) that puppet labs does
+    # not provide .deb for, then don't retrieve/install the package from puppet labs
+    arch=`uname -i`
+    if [ $arch != 'aarch64' ] ; then
+        puppet_deb=puppetlabs-release-${lsbdistcodename}.deb
+        wget http://apt.puppetlabs.com/$puppet_deb -O $puppet_deb
+        dpkg -i $puppet_deb
+        rm $puppet_deb
+    fi
 
     apt-get update
     DEBIAN_FRONTEND=noninteractive apt-get --option 'Dpkg::Options::=--force-confold' \
